Add files via upload
This project is about creating a simple API using FastAPI to calculate the factorial of a number. It has an endpoint called /factorial/{starting_number} that computes the factorial using a while loop. If you input 0, it will return {"result": false}. This API is a great way to learn how to build web services with FastAPI and practice some basic Python coding.